pyproject.toml
[project]
name = "portcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=2.0.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-alembic>=3.0.0",
    "hypercorn>=0.16.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
